requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.14.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
openapi-codec==1.3.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
selenium==4.9.1
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.4.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
wsproto==1.2.0
